ServerlessLambdas/requirements.txt
tinytag
passlib
pybranca
